% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sdc.R
\name{fit_sdc}
\alias{fit_sdc}
\title{Fit the Standard Diffusion Curve}
\usage{
fit_sdc(time_array, flux, n_max = 1, fixed_values = NULL, n_cores = NULL)
}
\arguments{
\item{time_array}{An array of time values in which the flux was collected.}

\item{flux}{A matrix or array of flux measurements.}

\item{n_max}{An integer for the total number of attempts for the LM-optimization.}

\item{fixed_values}{A list of fixed parameters. For example, assuming the flux is irreversible, then desorption must be zero, i.e., fixed_values = list(desorption = 0).}

\item{n_cores}{An integer for the total number cores to use in calculation.}
}
\value{
The standard diffusion curve based on the kinetic coefficients given.
}
\description{
Fit the standard diffusion curve coefficients given a flux (or a series of flux) and time array.
This function uses the Levenberg-Marquardt algorithm for a non-linear least squares fit of the coefficients.
}
\examples{
time_array = (1:5000) * 0.001

flux = sdc(time_array) + rnorm(length(time_array), mean = .2, sd = .01)
fit = fit_single_sdc(flux, time_array = time_array, fixed_values = list(desorption = 0))

flux2 = sdc(time_array) + rnorm(length(time_array), mean = .1, sd = .05)
flux = as.data.frame(cbind(flux, flux2))
names(flux) = c("f1", "f2")

fit = fit_sdc(time_array, flux, fixed_values = list(desorption = 0))
names(fit)
plot(fit$data$flux)
lines(fit$data$fitted, col = "red")

}
