% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_single_sdc.R
\name{fit_single_sdc}
\alias{fit_single_sdc}
\title{Fit the Standard Diffusion Curve}
\usage{
fit_single_sdc(time_array, flux, n_max = 1, fixed_values = NULL)
}
\arguments{
\item{time_array}{An array of time values in which the flux was collected.}

\item{flux}{An array of flux measurements.}

\item{n_max}{An integer for the total number of attempts for the LM-optimization.}

\item{fixed_values}{A list of fixed parameters. For example, assuming the flux is irreversible, then desorption must be zero, i.e., fixed_values = list(desorption = 0).}
}
\value{
A list of results including the time series data as a dataframe, the optimized coefficients as a dataframe, the non-linear least squares fit of the coefficients, and the least squares fit of the sdc to the original data.
}
\description{
Fit the standard diffusion curve coefficients given a flux and time array.
This function uses the Levenberg-Marquardt algorithm for a non-linear least squares fit of the coefficients.
}
\examples{
time_array = (1:5000) * 0.001
flux = sdc(time_array) + rnorm(length(time_array), mean = .2, sd = .01)
fit = fit_single_sdc(flux, time_array = time_array, fixed_values = list(desorption = 0))

plot(fit$data$time_array, fit$data$flux)
lines(fit$data$time_array, fit$data$fitted, col = "red")

}
