% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdc.R
\name{sdc}
\alias{sdc}
\title{Standard Diffusion Curve}
\usage{
sdc(time_array, diffusion = 1, adsorption = 0, desorption = 0, n_max = 2000)
}
\arguments{
\item{time_array}{An array of time values in which the flux was collected.}

\item{diffusion}{A scalar value of the diffusion coefficient.}

\item{adsorption}{A scalar value of the adsorption coefficient.}

\item{desorption}{A scalar value of the desorption coefficient.}

\item{n_max}{An integer for total number of indices in the iterative calculation of the SDC. Default is 2000 iterations.}
}
\value{
The standard diffusion curve based on the kinetic coefficients given.
}
\description{
Generate the standard diffusion curve from Gleaves et al TAP-2.
}
\details{
This function describes a theoretical 1-zone TAP flux including diffusion, adsorption, and desorption.
}
\examples{
# Recreation Figure 9 of Gleaves TAP-2

time_array = (1:5000) * 0.001
diffusion = 1
adsorption = 0
desorption = 0

flux1 = sdc(time_array, diffusion = diffusion,
  adsorption = adsorption, desorption = desorption
)

adsorption = 20
desorption = 20
flux2 = sdc(time_array, diffusion = diffusion,
  adsorption = adsorption, desorption = desorption
)

desorption = 5
flux3 = sdc(time_array, diffusion = diffusion,
  adsorption = adsorption, desorption = desorption
)

plot(time_array, flux1, type = "l", ylim = c(-0.1, 2.4),
xlab="Time", ylab = "Flux", main = "Black:ka =0, kd =0; Blue: ka = 20, kd = 20; Red: ka = 20, kd=5")
lines(time_array, flux2, col = "blue")
lines(time_array, flux3, col = "red")
}
